move virtual display to lib workspace
