+++ libs/virtual_display/Cargo.toml
@@ -3,18 +3,10 @@ name = "virtual_display"
 version = "0.1.0"
 edition = "2021"
 
-[lib]
-name = "virtual_display"
-crate-type = ["cdylib", "staticlib", "rlib"]
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[build-dependencies]
-cc = "1.0"
-
 [dependencies]
-thiserror = "1.0.30"
 lazy_static = "1.4"
-serde = "1.0"
-serde_derive = "1.0"
+libloading = "0.7"
 hbb_common = { path = "../hbb_common" }
+dylib_virtual_display = { path = "./dylib" }
